Refs 29380. Refactor MVP files and tidy code
