+++ scripts/Muon/GUI/ElementalAnalysis2/context/context.py
@@ -4,7 +4,7 @@
 #   NScD Oak Ridge National Laboratory, European Spallation Source,
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
-
+from typing import List
 from Muon.GUI.Common.muon_load_data import MuonLoadData
 
 
@@ -13,18 +13,17 @@ class DataContext(object):
         self.instrument = "rooth"
         self.current_runs = []
         self._loaded_data = load_data
-        self._run_info = {}
+        self._run_info = []
 
     @property
     def run_info(self):
         return self._run_info
 
-    def run_info_update(self, run, runObject):
-        self._run_info[run] = runObject
+    def run_info_update(self, run_object):
+        self._run_info.append(run_object)
 
     def clear_run_info(self):
-        for key in list(self._run_info.keys()):
-            del self._run_info[key]
+        self._run_info = []
 
 
 class ElementalAnalysisContext(object):
@@ -39,6 +38,6 @@ class ElementalAnalysisContext(object):
 
 class RunObject(object):
     def __init__(self, run, detectors, groupworkspace):
-        self._run_number = run
-        self._detectors = detectors
+        self._run_number: int = run
+        self._detectors: List[str] = detectors
         self._groupworkspace = groupworkspace
